Clone switcher controls into each block
- Allows on-hover reveal of switcher to happen reliably with CSS

- Requires refactoring of a few event-binding functions in order for
  them to operate on blocks rather than on a singleton node.

- Reliance on the `selectedBlock` global is diminished. The drawback is
a more hybrid code base, where some logic expects `selectedBlock` and
some doesn't need it. The advantage is that less mutating state needs to
be managed (setting `selectedBlock` can be delayed up until the point it
is necessary for a given operation).

- A couple of TypeErrors may pop up in the console. I suspect there are
some interactions where the target block isn't defined. Needs to be
addressed.