Introduce new CodeMirror block for editing code.
While demonstrating Gutenberg to my team the question was asked about
editing code blocks: "it's so terribly difficult in the WordPress
editor."

Wanting to demonstrate some of the flexibility that the block system
provides I created a new block type which turns a normal block into an
instance of the CodeMirror editor. This code block is like the freeform
block except that it transforms its content when loading into and saving
from the editor. In this case, I've pulled in the fairly robust `he`
library to perform HTML entity encoding. The result is that we should
never have to worry about issues with characters which are valid in code
but mess up HTML: `<`, `&`, etc...

Issues:
  - Yeah this pulls in CodeMirror directly
  - Yeah it assumes the language is JavaScript
  - Okay, it's a prototype...
  - Not sure if we're fine using `require()` the way I have